Editor Deprecation: Consolidate the editor URL selectors
Now that the only editor is Gutenberg, we are more concerned with
whether we should load the editor in the Gutenframe or redirect them to
`wp-admin`.

There were two selectors to get the appropriate editor URL,
`getGutenbergEditorUrl` and `getEditorUrl`. They were essentially the
same with subtle differences. This change consolidates them as
`getEditorUrl` and updates any references to `getGutenbergEditorUrl`.